feat: finished auth

+++ App/controllers/ListingController.php
@@ -3,7 +3,9 @@
 namespace App\Controllers;
 
 use Framework\Database;
+use Framework\Session;
 use Framework\Validation;
+use Framework\Authorization;
 
 class ListingController
 {
@@ -22,7 +24,7 @@ class ListingController
      */
     public function index()
     {
-        $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+        $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
 
         loadView('listings/index', [
             'listings' => $listings
@@ -80,7 +82,7 @@ class ListingController
         ];
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-        $newListingData['user_id'] = 1;
+        $newListingData['user_id'] = Session::get('user')['id'];
 
         $newListingData = array_map('sanitize', $newListingData);
 
@@ -101,10 +103,6 @@ class ListingController
                 'listing' => $newListingData
             ]);
         } else {
-            // $this->db->query('INSERT INTO listings(title, description, salary, tags, company, address,
-            //  city, state, phone, email, requirements, benefits, user_id) VALUES (:title, :description, :salary,
-            //  :tags, :address, :city, :state, :phone, :email, :requirements, :benefits, :user_id)', $newListingData);
-
             $fields = [];
 
             foreach ($newListingData as $field => $value) {
@@ -126,6 +124,8 @@ class ListingController
             $query = "INSERT INTO listings ({$fields}) VALUES ($values)";
             $this->db->query($query, $newListingData);
 
+            Session::setFlashMessage('success_message', 'Listing created successfully!');
+
             redirect('/listings');
         }
     }
@@ -146,16 +146,22 @@ class ListingController
         ];
 
         $listing = $this->db->query('SELECT * FROM listings WHERE id = :id', $params);
-
+        // Check if listing exists
         if (!$listing) {
             ErrorController::notFound('Listing not found!');
             return;
         }
 
+        // Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessage('error_message', 'You are not authorized to delete this listing');
+            return redirect('/listings/' . $listing->id);
+        }
+
         $this->db->query('DELETE FROM listings WHERE id = :id', $params);
 
         // Set flash message
-        $_SESSION['success_message'] = 'Listing deleted successfully!';
+        Session::setFlashMessage('success_message', 'Listing deleted successfully!');
 
         redirect('/listings');
     }
@@ -180,6 +186,12 @@ class ListingController
         if (!$listing) {
             ErrorController::notFound('Listing not found');
             return;
+        }
+
+        // Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessage('error_message', 'You are not authorized to update this listing');
+            return redirect('/listings/' . $listing->id);
         }
 
         loadView('listings/edit', [
@@ -207,6 +219,12 @@ class ListingController
         if (!$listing) {
             ErrorController::notFound('Listing not found');
             return;
+        }
+
+        // Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessage('error_message', 'You are not authorized to update this listing');
+            return redirect('/listings/' . $listing->id);
         }
 
         $allowedFields = [
@@ -249,7 +267,7 @@ class ListingController
             $updateValues['id'] = $id;
             $this->db->query($updateQuery, $updateValues);
 
-            $_SESSION['success_message'] = 'Listing Updated!';
+            Session::setFlashMessage('success_message', 'Listing Updated!');
 
             redirect('/listings/' . $id);
         }
